Update patch set 2
Patch Set 2: Code-Review-1

I'd like to handle this similar to the Krait bionic tweaks patchset - Publish the patchset somewhere it's easy to cherry-pick each individual patch, and put the squashed commit here for review.  As Guillaume said, please include the commit messages for each squashed patch in this one.

This will be used for review/test, and when it is determined to be OK, we'll push the individual commits that make up the squash and abandon this.

Patch-set: 2
Reviewer: Gerrit User 1000008 <1000008@56841c99-7c94-43d4-bf0b-cb4707e66d55>
Label: Code-Review=-1